Edu_backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0                  # async MongoDB driver
pymongo==4.7.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1                 # ← pin to 4.0.1 — passlib 1.7.4 is incompatible with bcrypt 4.1+
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
sentence-transformers==2.7.0
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
pymupdf==1.24.5               # fitz – PDF extraction
python-pptx==0.6.23           # PPTX/PPT extraction
python-docx==1.1.2            # DOCX/DOC extraction
httpx==0.27.0                 # async HTTP for Ollama calls
openpyxl==3.1.4               # Excel template generation & parsing